Modify bootstrap to skip chef install if already installed

+++ bootstrap.sh
@@ -10,6 +10,17 @@
 # 2020-02-04Bug fix on knife node environment set..
 # 2020-12-02 Various improvements
 
+CHEF_SEEN=false
+function check_chef {
+
+  rpm -q chef > /dev/null 2>&1
+  if [[ $? == 0 ]]
+    CHEF_SEEN=true
+  else
+    CHEF_SEEN=false
+  fi
+}
+
 # Ensure that the user runs this script (and the embedded
 # chef-client -z command from within the 'station' directory
 # this will ensure that chef zero creates a persistent 'nodes' directory
@@ -24,30 +35,35 @@ fi
 # link on Chef's download page in order to do so.  Until then, take the user
 # to this page and have them select it manually
 
-echo
-firefox --new-window https://downloads.chef.io/chef-workstation &
-echo
-
-echo
-echo "Manually download the latest Chef Workstation via Firefox"
-echo
-read -p "Press 'ENTER' when ready to proceed:"
-echo
-
-DL=/home/$USERNAME/Downloads
-
-ls -l $DL/chef*rpm
-CHEF_COUNT=$(ls -l $DL/chef*rpm | wc -l)
-if [[ $CHEF_COUNT -eq 1 ]]
+check_chef
+if [[ $CHEF_SEEN == false ]]
 then
-  sudo dnf -y install $DL/chef*rpm
-elif [[ $CHEF_COUNT -gt 1 ]]
-  echo "Error: I see more than one chef*rpm file in $DL"
-  echo "I don't know which one to install."
-  exit 1
-else
-  echo "Error: I can't find a $DL/chef*rpm file to install"
-  exit 1
+
+  echo
+  firefox --new-window https://downloads.chef.io/chef-workstation &
+  echo
+
+  echo
+  echo "Manually download the latest Chef Workstation via Firefox"
+  echo
+  read -p "Press 'ENTER' when ready to proceed:"
+  echo
+
+  DL=/home/$USERNAME/Downloads
+
+  ls -l $DL/chef*rpm
+  CHEF_COUNT=$(ls -l $DL/chef*rpm | wc -l)
+  if [[ $CHEF_COUNT -eq 1 ]]
+  then
+    sudo dnf -y install $DL/chef*rpm
+  elif [[ $CHEF_COUNT -gt 1 ]]
+    echo "Error: I see more than one chef*rpm file in $DL"
+    echo "I don't know which one to install."
+    exit 1
+  else
+    echo "Error: I can't find a $DL/chef*rpm file to install"
+    exit 1
+  fi
 fi
 
 # First run creates the chef-zero node object
